+++ ManejadorMetodos/CDBSicPro/Cpr_cuotapoliza.cs
@@ -195,7 +195,7 @@ namespace ManejadorMetodos.CDBSicPro
 
         }
 
-        public GetBuscarCuotaPolizaC_Result BuscarCuotaPolizaC(int id_poliza, int id_mov, int cuota)
+        public GetBuscarCuotaPolizaC_Result BuscarCuotaPolizaC(long id_poliza, long id_mov, int cuota)
         {
             try
             {
